@@ -11,7 +11,7 @@ namespace SolutionBuilder.WinForms
             // To customize application configuration such as set high DPI settings or default font,
             // see https://aka.ms/applicationconfiguration.
             ApplicationConfiguration.Initialize();
-            Application.Run(new frmMain());
+            Application.Run(new FrmMain());
         }
     }
 }
