@@ -26,7 +26,7 @@ namespace SolutionBuilder
         public override void WriteJson(JsonWriter writer, object? value, JsonSerializer serializer) => throw new NotImplementedException();
         public override object? ReadJson(JsonReader reader, Type objectType, object? existingValue, JsonSerializer serializer)
         {
-            List<Command?> commands = new();
+            List<Command?> commands = [];
             JArray jo = JArray.Load(reader);
 
             for (int i = 0; i < jo.Count; i++)
