@@ -6,7 +6,7 @@ namespace SolutionBuilder.Model
 {
     public class CommandReplaceCsProject : Command
     {
-        static readonly Dictionary<string, Project> ProjectsPath = new();
+        static readonly Dictionary<string, Project> ProjectsPath = [];
         public override bool Execute(SolutionFile solution, string _, Spectre.Console.Progress? progress = null)
         {
 
@@ -20,8 +20,10 @@ namespace SolutionBuilder.Model
                 {
                     if (!ProjectsPath.TryGetValue(projectInSolution.ProjectName, out project))
                     {
-                        ProjectOptions options = new();
-                        options.LoadSettings = ProjectLoadSettings.IgnoreMissingImports;
+                        ProjectOptions options = new()
+                        {
+                            LoadSettings = ProjectLoadSettings.IgnoreMissingImports
+                        };
 
                         project = Project.FromFile(projectInSolution.AbsolutePath, options);
                         ProjectsPath.Add(projectInSolution.ProjectName, project);
